# openapi.yaml
openapi: 3.1.0
info:
  title: SEMD API
  version: 1.0.0
paths:
...
  /auth/users:
    get:
      tags:
      - admin-users
      summary: '[Admin] Get All Users'
      description: Get all users with pagination. Admin only.
      operationId: get_all_users_auth_users_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number (1-indexed)
          default: 1
          title: Page
        description: Page number (1-indexed)
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Number of items per page
          default: 20
          title: Page Size
        description: Number of items per page
      - name: sort_by
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Field to sort by
          title: Sort By
        description: Field to sort by
      - name: sort_order
        in: query
        required: false
        schema:
          type: string
          pattern: ^(asc|desc)$
          description: Sort order (asc or desc)
          default: desc
          title: Sort Order
        description: Sort order (asc or desc)
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation error
    post:
      tags:
      - admin-users
      summary: '[Admin] Create User'
      description: Create a new user with password. Admin can create MEMBER users,
        Master Admin can create ADMIN users.
      operationId: create_user_auth_users_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateUserRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation error
  /auth/users/{user_id}:
    get:
      tags:
      - admin-users
      summary: '[Admin] Get User by ID'
      description: Get user details by ID. Admin only.
      operationId: get_user_by_id_auth_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation error
    put:
      tags:
      - admin-users
      summary: '[Admin] Update User'
      description: Update user details. Master Admin can change roles, Admin cannot.
      operationId: update_user_auth_users__user_id__put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation error
    delete:
      tags:
      - admin-users
      summary: '[Admin] Delete User'
      description: Delete user. Admin can delete MEMBER users, Master Admin can delete
        any user except other Master Admins.
      operationId: delete_user_auth_users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation error
...
  /dashboard/url-flag-stat:
    get:
      tags:
      - Dashboard
      summary: Geturlflagstat
      operationId: getUrlFlagStat_dashboard_url_flag_stat_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
...
  /dashboard/user-stat:
    get:
      tags:
      - Dashboard
      summary: Getuserstat
      operationId: getUserStat_dashboard_user_stat_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
...
  /setting/url-flag:
    get:
      tags:
      - url-flag
      summary: Get All URL Flags
      description: Get all URL flags. Requires authentication.
      operationId: get_all_flags_setting_url_flag_get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    post:
      tags:
      - url-flag
      summary: Create URL Flag
      description: Create a new URL flag. Requires authentication.
      operationId: create_flag_setting_url_flag_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlFlagCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
...
  /setting/url-flag/{flag_id}:
    get:
      tags:
      - url-flag
      summary: Get URL Flag by ID
      description: Get a specific URL flag by its ID.
      operationId: get_flag_by_id_setting_url_flag__flag_id__get
      parameters:
      - name: flag_id
        in: path
        required: true
        schema:
          type: integer
          title: Flag Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    put:
      tags:
      - url-flag
      summary: Update URL Flag
      description: Update an existing URL flag. Requires authentication.
      operationId: update_flag_setting_url_flag__flag_id__put
      parameters:
      - name: flag_id
        in: path
        required: true
        schema:
          type: integer
          title: Flag Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlFlagUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    delete:
      tags:
      - url-flag
      summary: Delete URL Flag
      description: Delete an existing URL flag. Requires authentication.
      operationId: delete_flag_setting_url_flag__flag_id__delete
      parameters:
      - name: flag_id
        in: path
        required: true
        schema:
          type: integer
          title: Flag Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
...
  /queue/url:
    get:
      tags:
      - queue
      summary: Get Retrain Queue
      description: Get all URLs waiting for retraining. Shows prediction info with
        user details.
      operationId: get_retrain_queue_queue_url_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /stat/user:
    get:
      tags:
      - stat-user
      summary: Get User Stat
      operationId: get_user_stat_stat_user_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/user/activity:
    get:
      tags:
      - stat-user
      summary: Get User Activity
      operationId: get_user_activity_stat_user_activity_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserActivityResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/user/role:
    get:
      tags:
      - stat-user
      summary: Get User Role
      operationId: get_user_role_stat_user_role_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/user/top:
    get:
      tags:
      - stat-user
      summary: Get Top Users
      operationId: get_top_users_stat_user_top_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopUserResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/api-key:
    get:
      tags:
      - stat-api-key
      summary: Get Api Key Stat
      operationId: get_api_key_stat_stat_api_key_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/api-key/usage:
    get:
      tags:
      - stat-api-key
      summary: Get Api Key Usage
      operationId: get_api_key_usage_stat_api_key_usage_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyUsageResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/api-key/trend:
    get:
      tags:
      - stat-api-key
      summary: Get Api Key Trend
      operationId: get_api_key_trend_stat_api_key_trend_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyTrendResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/api-key/endpoint:
    get:
      tags:
      - stat-api-key
      summary: Get Api Endpoint Stat
      operationId: get_api_endpoint_stat_stat_api_key_endpoint_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEndpointStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
...
  /stat/url-flag:
    get:
      tags:
      - stat-url-flag
      summary: Get Url Flag Stat
      operationId: get_url_flag_stat_stat_url_flag_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagStatResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/url-flag/trend:
    get:
      tags:
      - stat-url-flag
      summary: Get Url Flag Trend
      operationId: get_url_flag_trend_stat_url_flag_trend_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagTrendResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/url-flag/detail:
    get:
      tags:
      - stat-url-flag
      summary: Get Url Flag Detail
      operationId: get_url_flag_detail_stat_url_flag_detail_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagDetailResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
  /stat/url-flag/category:
    get:
      tags:
      - stat-url-flag
      summary: Get Url Flag Category
      operationId: get_url_flag_category_stat_url_flag_category_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlFlagCategoryResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
components:
  schemas:
...
    AdminCreateUserRequest:
      properties:
        username:
          type: string
          maxLength: 32
          minLength: 3
          title: Username
          description: Username
        email:
          type: string
          format: email
          title: Email
          description: Email address
        full_name:
          type: string
          maxLength: 512
          minLength: 1
          title: Full Name
          description: Full name
        password:
          type: string
          minLength: 8
          title: Password
          description: Password
        role:
          $ref: '#/components/schemas/RoleType'
          description: User role
          default: MEMBER
        birthday:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Birthday
          description: Birthday
        profile_img_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile Img Uri
          description: Profile image URI
      type: object
      required:
      - username
      - email
      - full_name
      - password
      title: AdminCreateUserRequest
...
    AdminUpdateUserRequest:
      properties:
        username:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 3
          - type: 'null'
          title: Username
          description: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
          description: Email address
        full_name:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Full Name
          description: Full name
        birthday:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Birthday
          description: Birthday
        profile_img_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile Img Uri
          description: Profile image URI
        role:
          anyOf:
          - $ref: '#/components/schemas/RoleType'
          - type: 'null'
          description: User role (Master Admin only)
      type: object
      title: AdminUpdateUserRequest
...
    ApiEndpointStatResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/ApiEndpointStatItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: ApiEndpointStatResponse
...
    ApiKeyStatResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          $ref: '#/components/schemas/ApiKeyStatItem'
      type: object
      required:
      - status
      - message
      - data
      title: ApiKeyStatResponse
...
    ApiKeyTrendResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/ApiKeyTrendItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: ApiKeyTrendResponse
...
    ApiKeyUsageResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/ApiKeyUsageItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: ApiKeyUsageResponse
...
    BaseResponseModel:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
      type: object
      required:
      - status
      - message
      title: BaseResponseModel
...
    FlagType:
      type: string
      enum:
      - MALICIOUS
      - BENIGN
      title: FlagType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueueListResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          additionalProperties: true
          type: object
          title: data
          description: Queue data with total and items
      type: object
      required:
      - status
      - message
      - data
      title: QueueListResponse
...
    RoleType:
      type: string
      enum:
      - GUEST
      - MEMBER
      - ADMIN
      - SUPER_ADMIN
      title: RoleType
...
    TopUserResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/TopUserItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: TopUserResponse
...
    UrlFlagCreateRequest:
      properties:
        url:
          type: string
          title: Url
          description: URL to flag
        type:
          $ref: '#/components/schemas/FlagType'
          description: Flag type (MALICIOUS or BENIGN)
          default: BENIGN
        access_level:
          $ref: '#/components/schemas/ACLType'
          description: Access level (GLOBAL or PRIVATE)
          default: PRIVATE
      type: object
      required:
      - url
      title: UrlFlagCreateRequest
...
    UrlFlagListResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items: {}
          type: array
          title: data
          description: List of URL Flags
      type: object
      required:
      - status
      - message
      - data
      title: UrlFlagListResponse
    UrlFlagResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          title: data
          description: URL Flag data
      type: object
      required:
      - status
      - message
      - data
      title: UrlFlagResponse
...
    UrlFlagStatResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          $ref: '#/components/schemas/UrlFlagStatItem'
      type: object
      required:
      - status
      - message
      - data
      title: UrlFlagStatResponse
...
    UrlFlagUpdateRequest:
      properties:
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
          description: URL to flag
        type:
          anyOf:
          - $ref: '#/components/schemas/FlagType'
          - type: 'null'
          description: Flag type (MALICIOUS or BENIGN)
        access_level:
          anyOf:
          - $ref: '#/components/schemas/ACLType'
          - type: 'null'
          description: Access level (GLOBAL or PRIVATE)
      type: object
      title: UrlFlagUpdateRequest
...
    UserActivityResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/UserActivityItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: UserActivityResponse
    UserListResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/UserModel'
          type: array
          title: data
          description: List of users
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
          title: pagination
          description: Pagination metadata
      type: object
      required:
      - status
      - message
      - data
      - pagination
      title: UserListResponse
...
    UserResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          $ref: '#/components/schemas/UserModel'
          title: data
          description: User data
      type: object
      required:
      - status
      - message
      - data
      title: UserResponse
...
    UserRoleStatResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/UserRoleStatItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: UserRoleStatResponse
...
    UserStatResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          $ref: '#/components/schemas/UserStatItem'
      type: object
      required:
      - status
      - message
      - data
      title: UserStatResponse
...